LogScriptEngine: move to cooja package
This moves the LogScriptEngine to the cooja
package and its interfaces to cooja.script.

This will enable the simulation and script
engine to communicate directly without
switching threads and without requiring
synchronization.
